move news_feed into content_management
