+++ src/main/content_management/dispatcher.ts
@@ -1,7 +1,7 @@
 import { fetchCurrentLocation } from "../fetch_cross_origin/ip_api";
 import { fetchUpcomingMeleeMajors } from "../fetch_cross_origin/melee_majors";
 import { fetchNearestTournaments } from "../fetch_cross_origin/smash_map";
-import { fetchNewsFeedData } from "../news_feed";
+import { fetchNewsFeedData } from "./news_feed";
 import type { ContentManagementParams, ContentManagementServiceName } from "./registry";
 
 type HandlerMap = {
