+++ src/main/content_management/news_feed.ts
@@ -3,9 +3,9 @@ import type { NewsItem } from "@common/types";
 import electronLog from "electron-log";
 import TurndownService from "turndown";
 
-import { getBlueskyFeed } from "./fetch_cross_origin/bluesky";
-import { getAllReleases } from "./fetch_cross_origin/github";
-import { getMediumFeed } from "./fetch_cross_origin/medium";
+import { getBlueskyFeed } from "../fetch_cross_origin/bluesky";
+import { getAllReleases } from "../fetch_cross_origin/github";
+import { getMediumFeed } from "../fetch_cross_origin/medium";
 
 const log = electronLog.scope("news_feed");
